LayoutTests/imported/w3c/web-platform-tests/IndexedDB/structured-clone.any.worker-expected.txt

CONSOLE MESSAGE: line 98: SyntaxError: No identifiers allowed directly after numeric literal

FAIL Indexed DB and Structured Serializing/Deserializing SyntaxError: No identifiers allowed directly after numeric literal
